requirements.txt
dnspython==2.3.0
PyQt5==5.15.9
PyQt5_sip==12.12.1
